backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1

# GitHub Integration
PyGithub==2.6.0

# AI/LLM Integration
openai==1.59.8
anthropic==0.43.2

# FREE tier LLM options (optional - install as needed)
google-generativeai==0.8.3  # Google Gemini (FREE)
groq==0.11.0                 # Groq (FREE & Fast)

# Testing
pytest==8.3.4
pytest-cov==6.0.0
httpx==0.28.1

# Development
python-dotenv==1.0.1